pg migrations int -> big int

+++ migrations/20250801121348_create_category_table.sql
@@ -2,9 +2,9 @@
 -- +goose StatementBegin
 SELECT 'up SQL query';
 CREATE TABLE category(
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(100) UNIQUE NOT NULL,
-    domain_id INT NOT NULL REFERENCES domain(id) ON DELETE CASCADE,
+    domain_id BIGINT NOT NULL REFERENCES domain(id) ON DELETE CASCADE,
     created_at TIMESTAMP NOT NULL DEFAULT now(),
     updated_at TIMESTAMP
 );
